More name refactoring from Glyph to Sequence. There may be more that I missed.
Boarder now has a more complete keyboard representation. It looks like both TAB and the KP_SEPARATOR are currently out of sorts. (It is odd that the KP_SEPARATOR used to work -- I think it broke with a Java update.) A few things on that screen need finished, but it is looking pretty good.

Changed from RESIZED_EVENT to RESIZE_EVENT. Ncurses calls it KEY_RESIZE so it is more consistent.

Fixed the issue with the keyboard in Boarder not getting all the colors from the palette -- the palette wasn't defined. :/ Easy enough to fix. The palette class now owns the function accepting an index or color and returning the color.