Some code-cleanup.
We now publicly state the code is under the GPL at the top of every file (I think -- I need to double-check).

We now require comments on all the code, so a lot of comments have been added. Most of them should be useful.

We're officially using Java properties for localization, so we've marked the non-localized strings. One example script has had the strings pulled out for localization.

A number of other improvements and tweaks came along with this update. There still a minor amount of code cleanup/restructuring to perform.