While creating a test-case for ColorNames, I ran in to some issues.
I also got rid of the functions in that class, and made the data contained in it static content. (The tests remain, as they also test the data.) This happened after the functions that existed there weren't sufficient for the needs of the test class and as such wouldn't be sufficient for normal use.

Updated the example scripts to use the new access to the named colors.

This points to a strong need for a specific JUnit test for ListMap.