Update for Issue 3.
This adds the TerminalInterface functions to support FullScreen mode.

There's support for a generalized Blacken-globlal hotkey to switch to FullScreen mode and application-level support to disable that functionality.

I need to diagnose an issue with Alt-Enter not working for the hotkey. For initial testing I've set it to use F11.