This isn't actively broken, but still needs quite a bit of polish.
That is, it does _something_ but not exactly the _correct_ something.

The display redraw logic still needs updated for the new cell contents, so we're getting highlights showcasing what is not right.

The way I'm placing the items doesn't respect doors, so they can lock off rooms and other number points.

Corridor placement is done before room placement, as a test this seems to showcase how much that is just all messed up. We see a lot of corridor/room:wall overlap that makes things weird.

Still for a "give me something using the BSP trees" proof of concept this seems to work pretty well. The BSP Tree side seems to be functioning fine, etc.