The library just got a few features. This adds support to Colors to test them.
I meant to add gradient support when I went through and took care of the rest of the LibTCOD color functions. At the time my mind was stuck on creating an efficient implementation that it glossed over creating a simple and functional implementation. Since gradients are going to be useful for the tutorial project, this adds gradient support.

I saw that SquidLib has "The Traditional Colors of Japan" as color options, and I desired more. We now have our first color set that will be distributed with Blacken as a resource file. Because our color names are looked up in a Map, and support alternate names, they may be referenced by their Japanese Kanji names, their Japanese "Romanized" names, the English names, and (sometimes) variations on the English names. Some of the names of colors are really awesome, like, "Brown noser's brown" (which has the alternate of "brown of flattery"), "Underside of willow leaves", "Rusty storeroom", "Brown rat grey", etc.