This finishes the image support.
Standard AWT images also support a .properties file -- so that means we can expose that file directly in the API without being weird with regards to text-images and graphics-images.

Standard images need a .properties file for two reasons (currently):
TRANSPARENT : specify the RGB color values which should be transparent (for BMP images without alpha or transparency information)
ALPHA : Specify the alpha channel setting below which the cell is transparent (Supports "*" as synonym for 255 and empty means only below 1)

Note that a "0" (zero) in the "ALPHA" setting means that the alpha channel is always provided and never used to indicate transparency. Since the rest of the UI currently has no layers to use the alpha channel that isn't going to hold much meaning for things.

The "TRANSPARENT" setting allows me to directly use a BMP file from the "rltiles" project in the example program. Since it is a BMP image there is no alpha channel or transparency information.