I think it resolves  50: Screen flickers with content for a moment then goes totally grey in OSX. (need to get final confirmation)
It appears to be faster than before I started working on this ticket.

The Mac OS X fix is in place -- but only for Mac OS X. (Though when in place on Windows it only decreased the speed by ~1-2 ms on the long cycles. -- Still under 10 ms on average for me.)

The Windows 7 fix has also been set to only apply to Windows 7 and it has been fixed so it doesn't visibly flicker the font as much.

Part of this fix included responding to more update events. What I'm checking in now doesn't ignore any repaint requests. This should clear up some edge conditions that were annoying.

The ReadOnlyAwtCell was put in place to help test some of the changes/fixes used for the speed improvement.