SingleLine.putString now has an optional "Alignment" enum that can alter the standard alignment. This is now favored over the "one-off" alignRight / centerText() functions that cropped up everywhere.
refresh(y, x) now _always_ refreshes the cell, ignoring the dirty status. This was normally how it was used already, I just needed to fix a couple refresh(void) calls so that they manually checked the dirty state.

Images has a function to refresh the cells based upon their color (foreground or background). This can be used when rotating the palette when we're not redrawing the screen after every move.

TerminalView has a concept of "Bounds" which includes a starting Y/X coordinates, but outside of the Bounds call it did not have a way to get that information. This has been corrected.