We now have a drippy texture that's used with a red "bloody" gradient for the Game Over screen.
GameOver takes some features from both the ConfirmationDialog as well as the HelpSystem. It allows for multiple tab-like pages of content and switching between them with left/right.

It can, however, be abstracted a little further as much of the same logic can be used to switch between races/classes while providing detailed documentation about the pros/cons of your particular choice.