This primarily adds an "ascension" mode to GameOver -- it is now possible to win the game.
This fixes the texture application on the "GameOver" page so that it uses more colors when the screen is taller. The idea is that the drip texture looks best when it has more than a screen full of indexes to play with.

Since the number of cells is now more variable, it became a major annoyance for me.

Initial sketch of a "player" object.