The Room and RoomFactory stuff has changed significantly.
We now have a "Renderable" interface which provides a method to get a TerminalCell for a thing. This has been extended in RenderableTimCell to take the TIMCell logic and provide a common means to get the renderable component. (As expected, Monster covers Item covers Terrain.)

We have a function to copy a Grid<Renderable> to a Grid<TerminalCell> in the DungeonUtils.

We have GridDungeon support. It's a dungeon based on a simple grid structure.

We have a "ConnectedRoom" which is based on a Room, but additionally can be "connected" to something. This is used with the standard RoomFactory via a method to change the class used for the Rooms. (Also, now all Room-types need a zero-argument constructor.)

There's some additional cleanup needed, but that will probably wait until after some additional refactoring in the tutorial module.